converted add page to dialog
